Change polling frequency from 5 minutes to 4 hours
A polling frequency of 5 minutes would sometimes cause multiple Jenkins
builds to stack, which was inconvenient.
As there is no hard requirement for builds to be triggered within
minutes of a change, this is a temporary change to reduce the clutter in
the Jenkins build page.
In the long-term, it is worth investigating whether setting up
post-commit hooks is a posibility, or whether it is possible to prevent
multiple builds from stacking at the same time. For now, however, this
change does the job.