Update Jenkins dependency to use my fork
As described in #351 in geerlingguy/ansible-role-jenkins, as a result of
Jenkins' migration from System V init to systemd, the Jenkins role is
now unable to override defaults that would have been previously set in
/etc/default/jenkins.
I would like to be able to rely on the playbook to set the
httpKeepAliveTimeout and httpsKeepAliveTimeout settings to higher
values, so that it is possible to download the AppImages without the
requests timing out after the default period of 5 seconds.
I plan to update my fork to use the new systemd init scripts used by the
latest versions of Jenkins, and then migrate this playbook back to using
the upstream geerlingguy's version if/when my or anyone else's chnage to
handle this gets merged.