Make create-user role an external dependency
Previously, this playbook included a role for creating a new linux user,
which was used by the playbook to create a user for the Jenkins
controller to ssh into on each individual agent.
I think it makes more sense to split out this role into a different
project, since creating a user has nothing to do with creating a Jenkins
server, and should not be tied to this project.
This commit makes use of the newly created project that this new role
now lives in.